# Conflicts:
#	pom.xml
#	src/main/java/shop/chaekmate/core/book/controller/BookController.java
#	src/main/java/shop/chaekmate/core/book/dto/request/BookCreateRequest.java
#	src/main/java/shop/chaekmate/core/book/dto/request/BookUpdateRequest.java
#	src/main/java/shop/chaekmate/core/book/repository/BookCategoryRepository.java
#	src/main/java/shop/chaekmate/core/book/repository/BookRepository.java
#	src/main/java/shop/chaekmate/core/book/service/BookService.java
#	src/main/java/shop/chaekmate/core/common/config/SwaggerConfig.java
#	src/main/java/shop/chaekmate/core/external/aladin/AladinSearchType.java
#	src/main/java/shop/chaekmate/core/external/aladin/controller/AdminAladinController.java
#	src/main/resources/application-dev.yml
#	src/test/java/shop/chaekmate/core/book/controller/CategoryControllerTest.java
#	src/test/java/shop/chaekmate/core/book/controller/LikeControllerTest.java
#	src/test/java/shop/chaekmate/core/book/controller/TagControllerTest.java

+++ src/main/java/shop/chaekmate/core/book/service/CategoryService.java
@@ -5,15 +5,25 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import shop.chaekmate.core.book.dto.request.CreateCategoryRequest;
+import shop.chaekmate.core.book.dto.response.CategoryHierarchyResponse;
 import shop.chaekmate.core.book.dto.response.CreateCategoryResponse;
 import shop.chaekmate.core.book.dto.response.ReadAllCategoriesResponse;
 import shop.chaekmate.core.book.dto.response.ReadCategoryResponse;
 import shop.chaekmate.core.book.dto.request.UpdateCategoryRequest;
 import shop.chaekmate.core.book.dto.response.UpdateCategoryResponse;
 import shop.chaekmate.core.book.entity.Category;
+import shop.chaekmate.core.book.exception.CategoryHasBookException;
+import shop.chaekmate.core.book.exception.CategoryHasChildException;
+import shop.chaekmate.core.book.exception.category.CategoryNotFoundException;
+import shop.chaekmate.core.book.exception.ParentCategoryNotFoundException;
+import shop.chaekmate.core.book.repository.BookCategoryRepository;
 import shop.chaekmate.core.book.repository.CategoryRepository;
 
 @Service
@@ -21,19 +31,16 @@ import shop.chaekmate.core.book.repository.CategoryRepository;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final BookCategoryRepository bookCategoryRepository;
 
     @Transactional
     public CreateCategoryResponse createCategory(CreateCategoryRequest request) {
-
-        if (request.name() == null) {
-            throw new RuntimeException("Category Name 이 Null 입니다.");
-        }
 
         Category parentCategory = null; // request.parentCategoryId 가 null 일때
 
         if (request.parentCategoryId() != null) {
             parentCategory = categoryRepository.findById(request.parentCategoryId())
-                    .orElseThrow(() -> new RuntimeException("해당하는 ID의 부모 카테고리를 찾을 수 없습니다."));
+                    .orElseThrow(ParentCategoryNotFoundException::new);
         }
 
         Category category = new Category(parentCategory, request.name());
@@ -54,7 +61,7 @@ public class CategoryService {
     public ReadCategoryResponse readCategory(Long targetCategoryId) {
 
         Category targetCategory = categoryRepository.findById(targetCategoryId)
-                .orElseThrow(() -> new RuntimeException("해당 ID의 카테고리를 찾을 수 없습니다"));
+                .orElseThrow(CategoryNotFoundException::new);
 
         String parentCategoryName = "null";
         if (targetCategory.getParentCategory() != null) {
@@ -69,12 +76,12 @@ public class CategoryService {
         Long parentCategoryId = request.parentCategoryId();
 
         Category targetCategory = categoryRepository.findById(targetId)
-                .orElseThrow(() -> new RuntimeException("해당 ID의 카테고리를 찾을 수 없습니다"));
+                .orElseThrow(CategoryNotFoundException::new);
 
         Category parentCategory = null;
         if (parentCategoryId != null) {
             parentCategory = categoryRepository.findById(parentCategoryId)
-                    .orElseThrow(() -> new RuntimeException("해당 ID의 부모 카테고리를 찾을 수 없습니다"));
+                    .orElseThrow(ParentCategoryNotFoundException::new);
         }
 
         targetCategory.updateCategory(parentCategory, request.name());
@@ -91,7 +98,17 @@ public class CategoryService {
     @Transactional
     public void deleteCategory(Long targetCategoryId) {
         Category targetCategory = categoryRepository.findById(targetCategoryId)
-                .orElseThrow(() -> new RuntimeException("해당 ID의 카테고리를 찾을 수 없습니다"));
+                .orElseThrow(CategoryNotFoundException::new);
+
+        if (bookCategoryRepository.existsByCategory(targetCategory)) {
+            throw new CategoryHasBookException();
+        }
+
+        if (categoryRepository.existsByParentCategory(targetCategory)) {
+            throw new CategoryHasChildException();
+        }
+
+        // TODO: 해당 카테고리에 해당하는 쿠폰 정책이 있을때 삭제불가
 
         categoryRepository.delete(targetCategory); // 실제론 deleted_at 이 바뀜
     }
@@ -118,5 +135,36 @@ public class CategoryService {
         }
 
         return roots;
+    }
+
+    @Transactional
+    public Page<CategoryHierarchyResponse> readAllCategoriesByPage(Pageable pageable) {
+        Page<Category> categoriesPage = categoryRepository.findAll(pageable);
+        List<Category> allCategories = categoryRepository.findAll();
+        Map<Long, String> categoryHierarchyMap = new HashMap<>();
+
+        for (Category category : allCategories) {
+            categoryHierarchyMap.put(category.getId(), buildCategoryHierarchy(category, allCategories));
+        }
+
+        List<CategoryHierarchyResponse> responses = categoriesPage.getContent().stream()
+                .map(category -> new CategoryHierarchyResponse(category.getId(),
+                        categoryHierarchyMap.get(category.getId())))
+                .toList();
+
+        return new PageImpl<>(responses, pageable, categoriesPage.getTotalElements());
+    }
+
+    private String buildCategoryHierarchy(Category category, List<Category> allCategories) {
+        StringBuilder hierarchy = new StringBuilder(category.getName());
+        Category current = category;
+        while (Objects.requireNonNull(current).getParentCategory() != null) {
+            Long parentId = current.getParentCategory().getId();
+            current = allCategories.stream().filter(c -> c.getId().equals(parentId)).findFirst().orElse(null);
+            if (current != null) {
+                hierarchy.insert(0, current.getName() + " > ");
+            }
+        }
+        return hierarchy.toString();
     }
 }
